requirements.txt
pytest
requests
PyYAML
allure-pytest
python-dotenv
jsonschema
pytest-dependency
pytest-ordering
